Function Used To Estimate Code Effects

This function estimates code effects using left and right embeddings from source and target sites.

Usage

CodeEff_Matrix(
  S.1,
  S.2,
  n1,
  n2,
  U.1,
  U.2,
  V.1,
  V.2,
  common_codes,
  zeta.int,
  lambda,
  p
)

Arguments

S.1

SPPMI from the source site.

S.2

SPPMI from the target site.

n1

The number of codes from the source site.

n2

The number of codes from the target site.

U.1

The left embeddings left singular vectors times the square root of the singular values from the source site.

U.2

The left embeddings left singular vectors times the square root of the singular values from the target site.

V.1

The right embeddings right singular vectors times the square root of the singular values from the source site.

V.2

The right embeddings right singular vectors times the square root of the singular values from the target site.

common_codes

The list of overlapping codes.

zeta.int

The initial estimator for the code effects.

lambda

The tuning parameter controls the intensity of penalization on the code effect.

p

The length of an embedding.

Value

A list with the following elements:

zeta

The estimated code effects.

dif_F

The Frobenius norm difference between the updated and initial estimators.

V.1.new

Updated right embeddings for the source site.

V.2.new

Updated right embeddings for the target site.