Function Used To Estimate Code-Site Effects Parallelly

Function Used To Estimate Code-Site Effects Parallelly

Usage

CodeSiteEff_l2_par(
  S.1,
  S.2,
  n1,
  n2,
  U.1,
  U.2,
  V.1,
  V.2,
  delta.int,
  lambda.delta,
  p,
  common_codes,
  n.common,
  n.core
)

Arguments

S.1

SPPMI from the source site

S.2

SPPMI from the target site

n1

the number of codes from the source site

n2

the number of codes from the target site

U.1

the left embeddings (left singular vectors times the square root of the singular values) from the source site

U.2

the left embeddings (left singular vectors times the square root of the singular values) from the target site

V.1

the right embeddings (right singular vectors times the square root of the singular values) from the source site

V.2

the right embeddings (right singular vectors times the square root of the singular values) from the target site

delta.int

the initial estimator for the code-site effect

lambda.delta

the tuning parameter controls the intensity of penalization on the code-site effects

p

the length of an embedding

common_codes

the list of overlapping codes

n.common

the number of overlapping codes

n.core

the number of cored used for parallel computation

Value

The output for the estimation of code-site effects