Function used to generate input data (used only for Simulations) Generate SPPMIs, dummy matrices based on prior group structures, and code-code pairs for tuning and evaluation

Function used to generate input data (used only for Simulations) Generate SPPMIs, dummy matrices based on prior group structures, and code-code pairs for tuning and evaluation

Usage

DataGen_rare_group(
  seed,
  p,
  n1,
  n2,
  n.common,
  n.group,
  sigma.eps.1,
  sigma.eps.2,
  ratio.delta,
  network.k,
  rho.beta,
  rho.U0,
  rho.delta,
  sigma.rare,
  n.rare,
  group.size
)

Arguments

seed

for reproducibility

p

the length of an embedding

n1

the number of codes in site 1

n2

the number of codes in site 2

n.common

common: the number of overlapping codes

n.group

the number of groups

sigma.eps.1

the sd of error in site 1

sigma.eps.2

the sd of error in site 2

ratio.delta

the proportion of codes in each site that have site-specific effects applied to them

network.k

the number of distinct blocks within each site for which unique inter-code correlations are modeled

rho.beta

AR parameter for the group effects covariance matrix

rho.U0

AR parameter for the code effects covariance matrix

rho.delta

AR parameter for the code-site effects covariance matrix

sigma.rare

the sd of error for rare codes (usually larger than sigma.eps.1 and sigma.eps.2)

n.rare

The number of rare codes

group.size

the size of each group

Value

Returns input data, SPPMIs, dummy matrices based on prior group structures and code-code pairs for tuning and evaluation