Function Used To Estimate Group Effects Parallelly

Function Used To Estimate Group Effects Parallelly

Usage

GroupEff_par(
  S.MGB,
  S.BCH,
  n.MGB,
  n.BCH,
  U.MGB,
  U.BCH,
  V.MGB,
  V.BCH,
  X.MGB.group,
  X.BCH.group,
  n.group,
  name.list,
  beta.int,
  lambda = 0,
  p,
  n.core
)

Arguments

S.MGB

SPPMI from the source site

S.BCH

SPPMI from the target site

n.MGB

the number of codes from the source site

n.BCH

the number of codes from the target site

U.MGB

the left embeddings (left singular vectors times the square root of the singular values) from the source site

U.BCH

the left embeddings (left singular vectors times the square root of the singular values) from the target site

V.MGB

the right embeddings (right singular vectors times the square root of the singular values) from the source site

V.BCH

the right embeddings (right singular vectors times the square root of the singular values) from the target site

X.MGB.group

the dummy matrix based on prior group structures at the source site

X.BCH.group

the dummy matrix based on prior group structures at the target site

n.group

the number of groups

name.list

the full list of code names from the source site and the target site with repeated names of overlapping codes

beta.int

the initial estimator for the group effects

lambda

the tuning parameter controls the intensity of penalization on the group effect; by default we set it to 0

p

the length of an embedding

n.core

the number of cored used for parallel computation

Value

The output of estimating group effects parallelly