Main function for MUGS algorithm

Main function for MUGS algorithm

Usage

MUGS(
  TUNE = F,
  Eva = T,
  Lambda = c(10),
  Lambda.delta = c(1000),
  n.core = 4,
  tol = 1,
  seed = 1,
  S.1 = NULL,
  S.2 = NULL,
  X.group.source = NULL,
  X.group.target = NULL,
  pairs.rel.CV = NULL,
  pairs.rel.EV = NULL,
  p = 100,
  n.group = 400
)

Arguments

TUNE

Logical value indicating whether the function should tune parameters 'TRUE' or use predefined parameters 'FALSE'.

Eva

Logical value indicating whether to perform evaluation (TRUE) or skip it (FALSE).

Lambda

The candidate values for the tuning parameter controls the intensity of penalization on the code effects.

Lambda.delta

The candidate values for the tuning parameter controls the intensity of penalization on the code-site effects.

n.core

Integer specifying the number of cores to use for parallel processing.

tol

Numeric value representing the tolerance level for convergence in the algorithm.

seed

Integer used to set the seed for random number generation, ensuring reproducibility of the simulated data or any stochastic process within the algorithm.

S.1

The SPPMI matrix from site 1.

S.2

The SPPMI matrix from site 2.

X.group.source

The dummy matrix on the group structure of codes at site 1.

X.group.target

The dummy matrix on the group structure of codes at site 2.

pairs.rel.CV

Code-code pairs used for tuning via cross validation

pairs.rel.EV

Code-code pairs used for evaluation

p

Integer indicating the length of embeddings.

n.group

The number of groups.

Value

The final result